requirements.txt
# Core DSPy and OpenAI
dspy-ai>=2.4.0
openai>=1.0.0

# Data processing
pandas>=2.0.0
numpy>=1.24.0
python-dotenv>=1.0.0

# File handling
python-docx>=0.8.11
PyPDF2>=3.0.0
openpyxl>=3.1.0

# Optional: For advanced text processing
transformers>=4.30.0
sentence-transformers>=2.2.0

# Development and testing
pytest>=7.0.0
jupyter>=1.0.0

# Utilities
tqdm>=4.65.0
rich>=13.0.0